pyproject.toml
[project]
name = "weather-api-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.6.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "jsonschema>=4.0.0",
]

[project.scripts]
weather-api-mcp-server = "weather_api_mcp_server.__main__:main"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["weather_api_mcp_server"]
# Ensure non-Python data is included in the wheel
force-include = { "weather_api_mcp_server/tools.json" = "weather_api_mcp_server/tools.json" }

[tool.hatch.build.targets.sdist]
include = [
  "weather_api_mcp_server/**",
  "README.md",
  "LICENSE",
]
